requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
sqlalchemy>=2.0
streamlit>=1.40
python-dotenv>=1.0
rank-bm25>=0.2
pandas>=2.2
